I finally caved and made the set up for an extern version of bheap
I didn't see an easy way to generate bheap_extern.h from bheap.h. I manually went in and deleted the function definitions-- not ideal.
bheap.h isn't that long so it was okay, but in the future, there should be a way to maybe mark the lines that are needed with comments or something and delete the other lines