I tried to make qsort use vla to benchmark
I ended up not being able to change qsort but qsort_r instead, and just passed NULL as the extra arg.
For the life of me, I could not figure out how __quicksort gets replaced with qsort, and not qsort_r when
__USE_GNU is defined before #include stdlib
anyways whatever. I don't think gcc is inlining the comparator calls, but theres too much asm to dig through for me